Use bzip2 -d instead of bunzip2
This is a PR that remedies the problem reported in #9. I've been informed that `bunzip2` is in fact an alias for `bzip2 -d`. Using the latter resolved the issue; it seemed the server had `bzip2` installed, but not the alias. Not sure if #9 is still relevant, feel free to close if not!